Move test method from JobScenariousTest to JobScenarioTest Delete JobScenariousTest.cs since it's redundant
