+++ SPMeta2/SPMeta2.Regression.Tests/Impl/Scenarios/JobScenariosTest.cs
@@ -1,20 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+using System.Collections.ObjectModel;
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SPMeta2.BuiltInDefinitions;
-using SPMeta2.Containers;
-using SPMeta2.Containers.Standard;
+
 using SPMeta2.Definitions;
-using SPMeta2.Enumerations;
-using SPMeta2.Models;
-using SPMeta2.Regression.Tests.Definitions;
 using SPMeta2.Regression.Tests.Impl.Scenarios.Base;
-using SPMeta2.Standard.Definitions;
 using SPMeta2.Syntax.Default;
-using SPMeta2.Validation.Validators.Relationships;
 
 namespace SPMeta2.Regression.Tests.Impl.Scenarios
 {
@@ -71,6 +62,36 @@ namespace SPMeta2.Regression.Tests.Impl.Scenarios
                 {
                     webApp.AddJob(jobDef);
                 });
+
+                TestModel(model);
+            });
+        }
+
+        [TestMethod]
+        [TestCategory("Regression.Scenarios.Job.Scopes")]
+        public void CanDeploy_Job_UnderWebApplication()
+        {
+            WithExpectedUnsupportedCSOMnO365RunnerExceptions(() =>
+            {
+                // OOTB job with 2 parameters for the constructor
+                var webAppJobDefinition = new JobDefinition
+                {
+                    Name = Rnd.String(),
+                    Title = Rnd.String(),
+                    ScheduleString = "yearly at jan 1 09:00:00",
+                    JobType = "Microsoft.SharePoint.Administration.SPDeadSiteDeleteJobDefinition, Microsoft.SharePoint",
+                    ConstructorParams = new Collection<JobDefinitionCtorParams>
+                    {
+                        JobDefinitionCtorParams.JobName,
+                        JobDefinitionCtorParams.WebApplication
+                    }
+                };
+
+                var model = SPMeta2Model
+                    .NewWebApplicationModel(webApp =>
+                    {
+                        webApp.AddJob(webAppJobDefinition);
+                    });
 
                 TestModel(model);
             });
